reintegration of ulong branch, plus some other changes accidentally checked in to that branch: ulong changes: - added stdint.h header for msvc because it isn't included by default, adds (u)int{8,16,32,64}_t types - changed "unsigned long" declarations in kernel to uint32_t when the 32-bit size was assumed and required, mainly hash values/ids and the functions that produce and use them - demonstrated increase in performance with 64-bit builds with learning on - updated retesave_{one_byte,two_bytes,four_bytes} functions to use the standard uint{8,16,32}_t types - changed "unsigned long long" to uint64_t misc changes: - rewrote a few inline functions so they better represent the macros that they replaced, some were introducing extra variables and casts that wouldn't be introduced if the macros were used, removed unnecessary casts - deallocate_instantiation algorithm changed to prevent a stack overflow triggered by retracting an instantiation with a huge chain of backtrace instantiations, see more comments in deallocate_instantiation, basically flattened out a recursive loop - added test to verify deallocate_instantiation change

+++ SoarKernel/include/agent.h
@@ -220,8 +220,8 @@ typedef struct agent_struct {
   
   
   /* Miscellaneous other stuff */
-  unsigned long       alpha_mem_id_counter; /* node id's for hashing */
-  unsigned long       beta_node_id_counter;
+  uint32_t       alpha_mem_id_counter; /* node id's for hashing */
+  uint32_t       beta_node_id_counter;
   struct ms_change_struct * ms_assertions;  /* changes to match set */
   struct ms_change_struct * ms_retractions;
 
@@ -283,7 +283,7 @@ typedef struct agent_struct {
   
   /* ----------------------- Symbol table stuff -------------------------- */
 
-  unsigned long       current_symbol_hash_id;
+  uint32_t       current_symbol_hash_id;
   unsigned long       id_counter[26]; 
   
   struct hash_table_struct * float_constant_hash_table;
